lesson 14 task 10

+++ src/main/java/lv/javaguru/java1/student_igor_eglit/lesson_14_project_property_insurance_calculator/PremiumCalculator.java
@@ -1,6 +1,8 @@
 package lv.javaguru.java1.student_igor_eglit.lesson_14_project_property_insurance_calculator;
 
+import java.math.BigDecimal;
+
 @FunctionalInterface
 interface PremiumCalculator {
-    float calculate(Policy policy);
+    BigDecimal calculate(Policy policy);
 }
